double to long

+++ rtp-core/src/main/java/io/github/dailystruggle/rtp/common/selection/region/selectors/memory/shapes/Circle.java
@@ -76,7 +76,7 @@ public class Circle extends MemoryShape<GenericMemoryShapeParams> {
   }
 
   @Override
-  public double xzToLocation(long x, long z) {
+  public long xzToLocation(long x, long z) {
     long cr = getNumber(GenericMemoryShapeParams.centerRadius, 64L).longValue();
     long cx = getNumber(GenericMemoryShapeParams.centerX, 0L).longValue();
     long cz = getNumber(GenericMemoryShapeParams.centerZ, 0L).longValue();
@@ -96,11 +96,11 @@ public class Circle extends MemoryShape<GenericMemoryShapeParams> {
 
     double radius = ((long) (Math.sqrt(x * x + z * z)));
 
-    return (radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI);
+    return (long) ((radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI));
   }
 
   @Override
-  public double xzToLocation(MutableRTPCoords coords) {
+  public long xzToLocation(MutableRTPCoords coords) {
     long cr = getNumber(GenericMemoryShapeParams.centerRadius, 64L).longValue();
     long cx = getNumber(GenericMemoryShapeParams.centerX, 0L).longValue();
     long cz = getNumber(GenericMemoryShapeParams.centerZ, 0L).longValue();
@@ -120,7 +120,7 @@ public class Circle extends MemoryShape<GenericMemoryShapeParams> {
 
     double radius = ((long) (Math.sqrt(x * x + z * z)));
 
-    return (radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI);
+    return (long) ((radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI));
   }
 
   @Override
